% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_button_serv.R
\name{dl_button_serv}
\alias{dl_button_serv}
\title{Handler de telechargement CSV}
\usage{
dl_button_serv(data, label)
}
\arguments{
\item{data}{Donnees a exporter.}

\item{label}{Nom du fichier (sans extension).}
}
\value{
Un objet downloadHandler.
}
\description{
Retourne un downloadHandler pour exporter des donnees en CSV.
}
