% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav_menu_apropos.R
\name{nav_menu_apropos}
\alias{nav_menu_apropos}
\title{Panneau de méthode, sources et crédits}
\usage{
nav_menu_apropos()
}
\value{
Un panneau de navigation.
}
\description{
Panneau de méthode, sources et crédits
}
