% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav_panel_exploration.R
\name{nav_panel_exploration}
\alias{nav_panel_exploration}
\title{Panneau principal d'exploration}
\usage{
nav_panel_exploration()
}
\value{
Le panneau de vue d'ensemble.
}
\description{
Panneau principal d'exploration
}
