% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Pretraitement des donnees CIR}
\usage{
preprocess(pathCIR = "data/data20-cir.xlsx")
}
\arguments{
\item{pathCIR}{Chemin vers le fichier Excel source.}
}
\value{
Invisible NULL. Le fichier Parquet est ecrit sur disque.
}
\description{
Charge un fichier Excel brut, nettoie les noms de colonnes
et sauvegarde en format Parquet.
}
