% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar_exploration.R
\name{sidebar_exploration}
\alias{sidebar_exploration}
\title{Barre latérale de filtres pour l'exploration}
\usage{
sidebar_exploration()
}
\value{
Les contrôles de filtre de l'application.
}
\description{
Barre latérale de filtres pour l'exploration
}
